[frontend] feat: add redirect
